Refactor method index from short to int in Redis metadata
Changed the type of method index from short to int in SpringRedisMetadataRepository and RedisCommandEventSerializer for improved compatibility and consistency. Updated serialization logic to use IntegerSerializer instead of ShortSerializer.

+++ microsphere-redis-spring/src/main/java/io/microsphere/redis/spring/metadata/SpringRedisMetadataRepository.java
@@ -149,7 +149,7 @@ public class SpringRedisMetadataRepository {
                 continue;
             }
 
-            short id = methodMetadata.getIndex();
+            int id = methodMetadata.getIndex();
             // Put id and Method as key
             if (redisMetadataCache.put(id, redisCommandMethod) == null && redisMetadataCache.put(redisCommandMethod, methodMetadata) == null) {
                 if (methodMetadata.isWrite()) {
@@ -198,12 +198,12 @@ public class SpringRedisMetadataRepository {
         return redisMetadata;
     }
 
-    public static Short findMethodIndex(Method redisCommandMethod) {
+    public static Integer findMethodIndex(Method redisCommandMethod) {
         MethodMetadata methodMetadata = (MethodMetadata) methodMetadataCache.get(redisCommandMethod);
         return methodMetadata == null ? null : methodMetadata.getIndex();
     }
 
-    public static Method findRedisCommandMethod(short methodIndex) {
+    public static Method findRedisCommandMethod(int methodIndex) {
         Method redisCommandMethod = (Method) methodMetadataCache.get(methodIndex);
         return redisCommandMethod;
     }
